% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_exp_gwp}
\alias{log_exp_gwp}
\title{Log-density for the expert assessment of GWP per-capita growth.}
\usage{
log_exp_gwp(model_out)
}
\arguments{
\item{model_out}{Data frame of model output (from \code{run_model}).}
}
\value{
Log-density value of the model average GWP per-capita growth rate.
}
\description{
\code{log_exp_gwp} returns the log-density for the expert assessment of
   GWP per-capita growth from Christensen et al (2018), which is used in
   probabilistic inversion.
}
\details{
This function uses the global per-capita growth trimmed mean (normal)
 distribution from 2010 to 2100 from Christensen et al (2018). It
 evaluates the density value of the model output based on this reference
 distribution.
}
